refactor: move DependencyResolverWorkspaceConfig to separate file
